renaming folder paths
