--- backend/src/main/kotlin/challenge/model/PaymentMethod.kt
@@ -1,3 +1,0 @@
-package challenge.model
-
-interface PaymentMethod
