+++ backend/src/main/kotlin/challenge/models/Address.kt
@@ -1,3 +1,3 @@
-package challenge.model
+package challenge.models
 
 data class Address(val street: String, val city: String, val country: String)
