+++ backend/src/main/kotlin/challenge/models/Customer.kt
@@ -1,3 +1,3 @@
-package challenge.model
+package challenge.models
 
 data class Customer(val name: String, val email: String)
