+++ backend/src/main/kotlin/challenge/models/Invoice.kt
@@ -1,4 +1,4 @@
-package challenge.model
+package challenge.models
 
 data class Invoice(val order: Order) {
     val billingAddress: Address = order.address
