+++ backend/src/main/kotlin/challenge/models/OrderItem.kt
@@ -1,4 +1,4 @@
-package challenge.model
+package challenge.models
 
 data class OrderItem(val product: Product, val quantity: Int) {
     val total: Double get() = product.price * quantity
