+++ backend/src/main/kotlin/challenge/models/Payment.kt
@@ -1,4 +1,4 @@
-package challenge.model
+package challenge.models
 
 import java.util.*
