+++ backend/src/main/kotlin/challenge/models/PaymentMethod.kt
@@ -1,0 +1,3 @@
+package challenge.models
+
+interface PaymentMethod
