+++ backend/src/main/kotlin/challenge/models/Product.kt
@@ -1,3 +1,3 @@
-package challenge.model
+package challenge.models
 
 data class Product(val name: String, val type: ProductType, val price: Double)
