+++ backend/src/main/kotlin/challenge/strategies/DigitalMediaPaymentProcessing.kt
@@ -1,6 +1,6 @@
 package challenge.strategies
 
-import challenge.model.Order
+import challenge.models.Order
 
 class DigitalMediaPaymentProcessing : PaymentProcessingStrategy {
     override fun processPayment(order: Order) {
