+++ backend/src/main/kotlin/challenge/strategies/MembershipPaymentProcessing.kt
@@ -1,6 +1,6 @@
 package challenge.strategies
 
-import challenge.model.Order
+import challenge.models.Order
 
 class MembershipPaymentProcessing : PaymentProcessingStrategy {
     override fun processPayment(order: Order) {
