+++ backend/src/test/kotlin/challenge/models/OrderTest.kt
@@ -1,0 +1,4 @@
+package challenge.models
+
+class OrderTest {
+}
