+++ backend/src/test/kotlin/challenge/strategies/BookPaymentProcessingTest.kt
@@ -1,0 +1,4 @@
+package challenge.strategies
+
+class BookPaymentProcessingTest {
+}
