+++ backend/src/test/kotlin/challenge/strategies/DigitalMediaPaymentProcessingTest.kt
@@ -1,0 +1,4 @@
+package challenge.strategies
+
+class DigitalMediaPaymentProcessingTest {
+}
