+++ backend/src/test/kotlin/challenge/strategies/MembershipPaymentProcessingTest.kt
@@ -1,0 +1,4 @@
+package challenge.strategies
+
+class MembershipPaymentProcessingTest {
+}
