+++ backend/src/test/kotlin/challenge/strategies/PhysicalProductPaymentProcessingTest.kt
@@ -1,0 +1,4 @@
+package challenge.strategies
+
+class PhysicalProductPaymentProcessingTest {
+}
